Add check and temporary fix for Postgres and transactions.
The query is blindly called without checking to see if the DB function exists.  This was causing issues in transactions, even though we were accounting for the issue client-side.  The reason this fix is temporary is due to it being 10-20% slower than before, and hopefully a faster solution can be found.