Add ability to upsert tables in schemas
When I tried to use this gem to upsert some tables in schemas in
Postgres, I found that it was throwing errors. I've patched this gem to
allow using tables in schemas. I have no real experience with the other
supported databases, so I'm not sure how these changes can (or should)
propogate to those other databases. For now, though, this allows me to
use this gem!